Make sure cacheops signals are attached first.
This fix makes sure the object is invalidated before all custom post_save signals are triggered.

It does this by directly manipulating a private Django API; the "receivers" attribute on the signals. This could be dangerous, but I only rely on it being a list, so this should work even over quite big changes to the API.