Skip test_148_reverse in Django 1.6-
It's no way to know if a model has proxy subclass with cacheops profile
until it's loaded. In Django 1.7+ all models are loaded before app ready
and hence cacheops install call so it's easy, but in older Djangos it's
hard to do properly.

For now I am not supporting this feature in Django 1.6-. And later I
will probably just drop old Djangos support.